chore: reorganize just a little
