+++ Cargo.toml
@@ -32,3 +32,4 @@ serde_json_path = "0.6.7"
 
 [dev-dependencies]
 rstest = "0.17.0"
+mockall = "0.12.*"
