+++ src/lib.rs
@@ -53,4 +53,6 @@ pub type Result<T, E = Error> = std::result::Result<T, E>;
 pub mod controller;
 
 mod mapping;
+mod resource_extensions;
 pub mod resources;
+mod util;
