+++ src/resource_extensions.rs
@@ -1,0 +1,33 @@
+use std::sync::Arc;
+
+use kube::{Api, ResourceExt};
+
+use crate::resources::ResourceSync;
+use crate::FINALIZER;
+
+impl ResourceSync {
+    pub(crate) fn has_been_deleted(&self) -> bool {
+        self.metadata.deletion_timestamp.is_some()
+    }
+
+    pub(crate) fn has_target_finalizer(&self) -> bool {
+        self.metadata
+            .finalizers
+            .as_ref()
+            .map_or(false, |f| f.contains(&FINALIZER.to_string()))
+    }
+
+    pub(crate) fn api(&self, ctx: Arc<crate::controller::Context>) -> Api<Self> {
+        match self.namespace() {
+            None => Api::all(ctx.client.clone()),
+            Some(ns) => Api::namespaced(ctx.client.clone(), &ns),
+        }
+    }
+
+    pub(crate) fn finalizers_clone_or_empty(&self) -> Vec<String> {
+        match self.metadata.finalizers.as_ref() {
+            Some(f) => f.clone(),
+            None => vec![],
+        }
+    }
+}
